add getCollectivityById and getFinancialAccounts

+++ Ex/src/main/java/com/example/Ex/DTO/CreateMembershipFeeDTO.java
@@ -1,19 +1,32 @@
 package com.example.Ex.DTO;
 
-import java.math.BigDecimal;
-
 public class CreateMembershipFeeDTO {
     private String eligibleFrom;
     private String frequency;
-    private BigDecimal amount;
+    private double amount;
     private String label;
 
-    public String getEligibleFrom() { return eligibleFrom; }
-    public void setEligibleFrom(String eligibleFrom) { this.eligibleFrom = eligibleFrom; }
-    public String getFrequency() { return frequency; }
-    public void setFrequency(String frequency) { this.frequency = frequency; }
-    public BigDecimal getAmount() { return amount; }
-    public void setAmount(BigDecimal amount) { this.amount = amount; }
-    public String getLabel() { return label; }
-    public void setLabel(String label) { this.label = label; }
+    public String getEligibleFrom() {
+        return eligibleFrom;
+    }
+    public void setEligibleFrom(String eligibleFrom) {
+        this.eligibleFrom = eligibleFrom; }
+    public String getFrequency() {
+        return frequency;
+    }
+    public void setFrequency(String frequency) {
+        this.frequency = frequency;
+    }
+    public double getAmount() {
+        return amount;
+    }
+    public void setAmount(double amount) {
+        this.amount = amount;
+    }
+    public String getLabel() {
+        return label;
+    }
+    public void setLabel(String label) {
+        this.label = label;
+    }
 }
